### 2019-12-19 #1963 @DorianDepriester
Dear all,
As potential reviewers, I suggest:

- streeve
- IgorBaratta
- ralfHielscher (not in the list of JOSS' reviewers)
- lheagy
- WalBytes

Regards.
