### 2019-12-20 b110680
-add extra details in help -first check if Gmsh is reachable (on PC)
