### 2019-12-20 #1963 @labarba
👋 @meg-simula — Could you handle this submission as editor?
The author suggested @Kevin-Mattheus-Moerman, but he is handling 5 papers right now.
Check out the [article proof.](https://github.com/openjournals/joss-papers/blob/joss.01963/joss.01963/10.21105.joss.01963.pdf)
