### 2019-12-20 04c97e6
Add comprehensive help sections to all methods and functions
