+++ gmshGeo.m
@@ -1,17 +1,20 @@
 classdef gmshGeo
     
     properties
-        V=[];
-        Segments=cell(0,1);
-        Grains=table;
-        SingularPoints=[];
-		Interfaces=struct;
+        V=[];               % Vertices (BSpline knots)
+        Segments=cell(0,1); % Lists of knots, defining the BSplines
+        Grains=table;       % Table summurizing the properties of each grain
+        SingularPoints=[];  % List of singular points (Triple junctions, corners etc.)
+		Interfaces=struct;  % Phase-to-phase interfaces
     end
     
     methods
         function G=gmshGeo(grains)
-        %Object constructor.
-        % Computes the gmshGeo object from a set of grains (grain2d object)
+        %GMSHGEO Object constructor.
+        % Computes the GMSHGEO object from a set of grains (grain2d
+        % object).
+		%
+		% See also savegeo, exportGrainProps
 			if ~isa(grains,'grain2d')
 				error('Input argument must be of class grain2d');
 			end
@@ -73,8 +76,13 @@ classdef gmshGeo
         end
         
         function plot(obj,varargin)
-        %Plots the segments found in each grains.
-        % PLOT(Gmsh) plots all the segments found.
+        %PLOT Plot the segments found in each grains.
+        % PLOT(Object) plots all the segments.
+		%
+		% PLOT(Object,I) with I an array of integers, plots only the grains
+		% whose indices belong to I.
+		%
+		% See also plotElementSize
  			if nargin>1
  				GrainIDs=varargin{1};
 				lineList=uint16(segmentList(obj,GrainIDs));
@@ -138,30 +146,40 @@ classdef gmshGeo
         end
         
         function savegeo(obj,filepath,varargin)
-        %Saves the geometry as an input file for Gmsh (*.geo).
-        %	- SAVEGEO(Gmsh,filepath) saves the geometry in the
-        % corresponding file path. The element size for meshing is equalt
-        % to the EBSD resolution.
-		%	- SAVEGEO(...,'ElementSize',value) results in element sizes
-		%  equal to the given value.
-		%	- SAVEGEO(...,'Thickness',value) sets an extrusion thickness
-		%  equal to the given value (equal to element size by default).
-        %	- SAVEGEO(...,'gradient',k) results in elements with size equal
-        %  to s+k*d (d being the distance from the nearest boundary and s 
-		%  the default element size).
-		%	- SAVEGEO(...,'ElementType',type) sets the element type used
-		%  for meshing. It can be:
+        %SAVEGEO Save the geometry as an input file for Gmsh (*.geo).
+        %	SAVEGEO(Object,filepath) saves the geometry in the
+        %	corresponding file path. The element size for meshing is equal
+        %	to the EBSD resolution.
+        %
+		%	SAVEGEO(...,'ElementSize',value) results in element sizes
+		%	equal to the given value.
+        %
+		%	SAVEGEO(...,'Thickness',value) sets an extrusion thickness
+		%	equal to the given value (equal to element size by default).
+        %
+        %	SAVEGEO(...,'gradient',k) results in elements with size equal
+        %	to s+k*d (d being the distance from the nearest boundary and s 
+		%	the default element size).
+        %
+		%	SAVEGEO(...,'ElementType',type) sets the element type used
+		%	for meshing. It can be:
 		%		-'Wedge' (default) for Wedge elements,
 		%		-'Brick' for quadrangular (2D)/Brick (3D) elements.
[…]
