### 2019-12-20 5625c89
Do not check if Gmsh is in the PATH (avoid ugly message)
