### 2019-12-21 #1963 @meg-simula
I would be happy to, but I am out of office until Mon Jan 6 2020. If it can
wait until then, feel free to assign me.

On Fri, Dec 20, 2019, 16:43 Lorena A. Barba <notifications@github.com>
wrote:

> 👋 @meg-simula <https://github.com/meg-simula> — Could you handle this
> submission as editor?
> The author suggested @Kevin-Mattheus-Moerman
> <https://github.com/Kevin-Mattheus-Moerman>, but he is handling 5 papers
> right now.
> Check out the article proof.
> <https://github.com/openjournals/joss-papers/blob/joss.01963/joss.01963/10.21105.joss.01963.pdf>
>
