### 2019-12-21 #1963 @labarba
Thanks, @meg-simula ! Yes, all editor assignments this week are carrying this caveat 😁
