### 2020-01-06 bef1bbd
use the public function EulerPath from MTEX instead of home-made function EulerCycles
