### 2020-01-06 0852f0f
Use MTEX EulerCycles instead
