### 2020-01-06 2be4e37
Update BibTeX entry
Update DOI and version
