### 2020-01-07 3605478
-computeSegments.m: remove unnecessary loop when looking for segment duplicates -gmshGeo.m: savegeo method: no longer multiply the segment indices by 10 (unnecessary for newer versions of Gmsh)
