### 2020-01-07 5064ee4
use the built-in writetable function instead of a for loop to export the grain properties
