### 2020-01-07 df27d0e
Add gathered method (mesh) for writing the geometry into a temporary file and meshing it with the Gmsh command. Add help section and update 'see also' Implementation of mesh() method into Example.m, adding extra comments for this shortcut
