+++ Gmsh.m
@@ -17,6 +17,7 @@ function varargout=Gmsh(inputFilePath,outputFilePath)
 %
 %   s=GMSH(...) saves the message a string.
 %
+%	See also gmshGeo/mesh gmshGeo/savegeo
 	
 	%% Remove file extension from input file
 	[inputPath,inputFileName,~] = fileparts(inputFilePath);
