### 2020-01-07 3833695
add complete help in the mesh() method
