### 2020-01-07 2f8e06e
Fix typos in help sections
