### 2020-01-07 5be6eeb
Fix error when path contains space(s)
