### 2020-01-07 de36883
Remove the Gmsh function and implement it directly in the mesh() method Update the Example.m accordingly
