### 2020-01-09 f97f672
fix the self loop error roll back to version 72, using home-made EulerPath.m, instead of public EulerCycle() from MTEX
