### 2020-01-09 2d2d096
Add linear indexing feature for selecting the grains by their index or by their phase name Simplify the plot method (extra argument non longer necessary) Add varargin option to the plot, passed to the built-in plot function Update help for the plot method add help for the constructor
