### 2020-01-09 db575aa
Implement the end method for the gmshGeo class
