### 2020-01-09 b01cef6
Define a specific function for inheriting the plotting orientation from MTEX Implementation of that function into the plot and plotElementSize methods Fix typos in help sections
