### 2020-01-09 0659845
change mutliple space to tab in help sections
