+++ gmshGeo.m
@@ -1,16 +1,16 @@
 classdef gmshGeo
     
     properties
-        V=[];               % Vertices (BSpline knots)
-        Segments=cell(0,1); % Lists of knots, defining the BSplines
-        Grains=table;       % Table summurizing the properties of each grain
-        SingularPoints=[];  % List of singular points (Triple junctions, corners etc.)
-		Interfaces=struct;  % Phase-to-phase interfaces
+		V=[];		       %	Vertices (BSpline knots)
+		Segments=cell(0,1); %	Lists of knots, defining the BSplines
+		Grains=table;       %	Table summurizing the properties of each grain
+		SingularPoints=[];  %	List of singular points (Triple junctions, corners etc.)
+		Interfaces=struct;  %	Phase-to-phase interfaces
     end
     
     methods
-        function G=gmshGeo(grains)
-        %GMSHGEO Object constructor.
+		function G=gmshGeo(grains)
+		%GMSHGEO Object constructor.
 		%
 		%	GMSHGEO(GRAINS) constructs an instance of class GMSHGEO, from
 		%	the object GRAINS (of class grain2d).
@@ -25,11 +25,11 @@ classdef gmshGeo
 		%	given phase. E.g.: obj('Forsterite') will keep only the data
 		%	related to the phase named Forsterite.
 		%
-		% See also calcGrains, mesh, exportGrainProps.
+		%	See also calcGrains, mesh, exportGrainProps.
 			if ~isa(grains,'grain2d')
 				error('Input argument must be of class grain2d');
 			end
-            [Segmts,OuterLoop,InnerLoops,G.V,G.SingularPoints]=computeSegments(grains);
+		    [Segmts,OuterLoop,InnerLoops,G.V,G.SingularPoints]=computeSegments(grains);
 			GrainID=grains.id;
 			phaseList=grains.mineralList;
 			Phase=phaseList(grains.phaseId)';
@@ -47,7 +47,7 @@ classdef gmshGeo
 			
 			
 			
-			% Initialize the list of phase-to-phase interfaces
+			%	Initialize the list of phase-to-phase interfaces
 			np=length(grains.mineralList);			
 			for i=1:np
 				for j=i:np
@@ -59,10 +59,10 @@ classdef gmshGeo
 					end
 				end
 			end
-			border='Border';	% Name for the domain boundary
+			border='Border';	%	Name for the domain boundary
 			GB.(border)=uint16([]);
 			
-			% Fetch the interfaces
+			%	Fetch the interfaces
 			for ids=1:size(Segmts,1)
 				ph2ph=Segmts{ids,2};
 				i=min(ph2ph);
@@ -77,18 +77,18 @@ classdef gmshGeo
 				GB.(strname)=[GB.(strname) ids];
 			end
 			
-			% Remove empty sets
+			%	Remove empty sets
 			fn = fieldnames(GB);
 			tf = cellfun(@(c) isempty(GB.(c)), fn);
 			
-			% Update properties
+			%	Update properties
 			G.Interfaces = rmfield(GB, fn(tf));
 			G.Segments=Segmts(:,1);
-        end
-        
-        function plot(obj,varargin)
-        %PLOT Plot the segments found in each grains.
-        %	PLOT(obj) plots all the segments.
+		end
+		
+		function plot(obj,varargin)
+		%PLOT Plot the segments found in each grains.
+		%	PLOT(obj) plots all the segments.
 		%
 		%	PLOT(obj(I)) plots the segments of grains whose indices are
 		%	given by the array I.
@@ -96,7 +96,7 @@ classdef gmshGeo
 		%	PLOT(obj(P)) plots the segments of grains of phase P only,
 		%	where P is a string.
 		%
[…]
