### 2020-01-09 c8eeb17
remove obsolete comment
