### 2020-01-13 #1963 @meg-simula
Hi @DorianDepriester, thank you for the submission! This looks nicely in scope for JOSS.

@lheagy @streeve @ralfHielscher Would you be willing and available to review this submission?
