### 2020-01-14 03d7bee
new method: smooth()
