### 2020-01-14 10dabcf
The plot() method no longer uses 'hold on/off', allowing to use it externally (eg. for superimposing grain2d/gmshGeo data)
