### 2020-01-15 00a023d
remove unused statement
