### 2020-01-15 c96e6e5
Update help section for the size() method
