### 2020-01-15 c43da5a
Remove unused vertices when applying the simplify() method
