### 2020-01-16 #1963 @streeve
Yes, I'm happy to - this will be my first JOSS review :)
