### 2020-01-16 8ad6dd0
Do not show the legend when plotting Fix legend issue when plotting with 'hold on'
