### 2020-01-17 3542936
The Interface attribute is now a table, allowing to name the interfaces with both spaces and parentheses
