### 2020-01-17 4e5ce3c
The singular points are computed from an external function (singularPoints.m) and the corner of the ROI are no longer needed.
