### 2020-01-17 6a1b2ce
smooth() method deprecated; replaced by the function cond_smooth.m for grains
