### 2020-01-17 b11ccb9
Add conditionnal smoothing of grains
