### 2020-01-17 05b6cce
add waitbar to the simplify function and update the list of Singular points
