### 2020-01-17 0213ee3
Vectorized way for updating the segments in the simplify() method
