### 2020-01-17 64cc6fd
fix issue because of transposed segments after applying the simplify() method keep datatype constant for the segment when applying the simplify() method
