### 2020-01-20 d9d6fff
Force Gmsh verbosity level to 4
