### 2020-01-20 13967e8
The mesh() command prompts a uigetfile for locating the Gmsh binary if Gmsh is not found
