+++ Uninstall_MTEX2Gmsh.m
@@ -1,0 +1,7 @@
+function Uninstall_MTEX2Gmsh
+if ispref('MTEX2Gmsh')
+	rmpref('MTEX2Gmsh');
+end
+fp=fileparts(mfilename('fullpath'));	% Path to the present file
+rmpath(fp);
+savepath;
