### 2020-01-23 8aa4e48
If the path to Gmsh has to be defined when running the msh() command, ask the user if (s)he wants to save the path permanently
