### 2020-01-27 3044eaf
Update MTEX2Gmsh version number to 2.0
