### 2020-01-31 #1963 @ralfHielscher
Hi Marie,

I also agree to review. What are the next steps?

Ralf.

********************************************************************
Ralf Hielscher                   Tel: +371-531-38556
Fakultät für Mathematik               +371-531-22200 (Sekr.)
Technische Universität Chemnitz  Fax: +371-531-22109
Reichenhainer Str. 39            E-mail:
ralf.hielscher@mathematik.tu-chemnitz.de
D-09126 Chemnitz                 http://www.tu-chemnitz.de/~rahi
********************************************************************

On Mon, Jan 13, 2020 at 2:46 PM Marie E. Rognes <notifications@github.com>
wrote:

> Hi @DorianDepriester <https://github.com/DorianDepriester>, thank you for
> the submission! This looks nicely in scope for JOSS.
>
> @lheagy <https://github.com/lheagy> @streeve <https://github.com/streeve>
> @ralfHielscher <https://github.com/ralfHielscher> Would you be willing
> and available to review this submission?
>
