### 2020-02-06 92692df
Delete the 'QuoteStrings' option in exportGrainProps() for backward compatibility with older versions of MATLAB
