+++ docs/readme.md
@@ -1,0 +1,2 @@
+# MTEX2Gmsh
+The toolbox for Matlab is intended to generate meshes from EBSD data, in order to perform Finite Element simulations at grain scale on crystalliune materials (such as minerals or metals)
