### 2020-02-10 807f3c2
Add metadata and logo to config file
