### 2020-02-10 27847c4
Remove redundant title and fix typo
