+++ docs/README.md
@@ -1,2 +1,1 @@
-# MTEX2Gmsh
-The toolbox for Matlab is intended to generate meshes from EBSD data, in order to perform Finite Element simulations at grain scale on crystalliune materials (such as minerals or metals)
+The toolbox for Matlab is intended to generate meshes from EBSD data, in order to perform Finite Element simulations at grain scale on crystalline materials (such as minerals or metals)
