### 2020-02-10 30413ab
Change to larger logo
