+++ docs/README.md
@@ -1,1 +1,1 @@
-The toolbox for Matlab is intended to generate meshes from EBSD data, in order to perform Finite Element simulations at grain scale on crystalline materials (such as minerals or metals)
+<iframe src="html/index.html" style="width: 500px; height: 130px; border: 0px"></iframe>
