+++ docs/README.md
@@ -1,1 +1,1 @@
-<iframe src="html/index.html" style="width: 500px; height: 130px; border: 0px"></iframe>
+<iframe src="html/index.html" style="width: 500px; height: 1000px; border: 0px"></iframe>
