+++ docs/README.md
@@ -1,1 +1,23 @@
-<iframe src="html/index.html" style="width: 500px; height: 1000px; border: 0px"></iframe>
+# Purpose of this toolbox
+In order to evaluate the thermo-mechanical behaviour of crystalline materials (such as metals or ceramics) at microscopic scale, one usually perform numerical simulation at grain scale using the Finite Element Method. In order to proceed, one must first create a mesh which is representative of the real material.
+
+The microstructure of cytalline materials is usually made from Electron Backscattered Diffraction (EBSD) technique. Thus, this toolbox is designed to generate meshes from EBSD in a robust and accurate way.
+
+# Examples
+
+
+# Documentation
+## Online
+You can navigate the documentation through the following sections:
+ - [Install/Uninstall](html/install.html)
+ - [Mesh](html/meshing.html)
+ - [Plot](html/plotting)
+ - [Customize the geometry](custom)
+ - [Improve the grain boundaries](improvegeometry)
+
+## From MATLAB
+Once installed on your Matlab session, open the documentation of the present toolbox by typing:
+
+    doc
+    
+Then, click on "MTEX2Gmsh toolbox", under the _Supplemental Software_ section (bottom right).
