+++ docs/README.md
@@ -12,8 +12,8 @@ You can navigate the documentation through the following sections:
  - [Install/Uninstall](html/install.html)
  - [Mesh](html/meshing.html)
  - [Plot](html/plotting)
- - [Customize the geometry](custom)
- - [Improve the grain boundaries](improvegeometry)
+ - [Customize the geometry](html/custom)
+ - [Improve the grain boundaries](html/improvegeometry)
 
 ## From MATLAB
 Once installed on your Matlab session, open the documentation of the present toolbox by typing:
