+++ docs/Examples/aachen.m
@@ -1,0 +1,6 @@
+mtexdata aachen
+ebsd=ebsd('indexed');
+grains=calcGrains(ebsd);
+G=gmshGeo(grains);
+G=simplify(G);
+mesh(G,'aachen.msh','ElementSize',0.7,'gradient',0.2)
