+++ docs/README.md
@@ -4,8 +4,12 @@ In order to evaluate the thermo-mechanical behaviour of crystalline materials (s
 The microstructure of cytalline materials is usually made from Electron Backscattered Diffraction (EBSD) technique. Thus, this toolbox is designed to generate meshes from EBSD in a robust and accurate way.
 
 # Examples
-[![Example: aachen.m](./Examples/aachen.png)](Examples/aachen.m)
-[![Example: titanium_medium.m](./Examples/titanium_medium.png)](Examples/titanium_medium.m)
+![Example: aachen.m](./Examples/aachen.png)
+[``aachen.m``](Examples/aachen.m)
+
+
+![Example: titanium_medium.m](./Examples/titanium_medium.png)
+[``titanium_medium.m``](Examples/titanium_medium.m)
 
 # Documentation
 ## Online
