### 2020-02-11 560ee79
Simplified version for script twins.m
