### 2020-02-11 5a08e34
Add superimposed example and add details about selection by phase name
