### 2020-02-11 66a109b
Remove Example.m (see docs/Examples/ instead)
