### 2020-02-19 DorianDepriester/MTEX2Gmsh#1 @streeve
Documentation
I have a few suggestions to improve documentation per openjournals/joss-reviews#2094

Concerning the tutorial:
- [x] Specify matlab, mtex, and gmsh versions required (potentially gcc underlying gmsh)
- [x] Data initialization should have a heading and code block formatting (currently only done for the "Improve" section), ideally pointing out this examples from mtex are being used
- [x] Link to doc page in README (maybe add image)
- [x] In the "Mesh" and "Customize" sections the figures are not directly from code examples - either be very clear about this or add how to obtain these images as part of the tutorial

The example cases are easy to follow and seem to cover the extent of the intended uses, but unit testing is missing. Some recently accepted joss papers using matlab have what seems to be an offline-automated test suite
- [x] Add unit tests

I did not see any API documentation (maybe I am forgetting how to access this in matlab)
- [x] Add API docs

Finally, make sure to add a CONTRIBUTING.md file for Github contributors
- [x] CONTRIBUTING.md
