### 2020-02-19 a6fd986
Add link to index.html into all documentation pages Add plain heading code for example dataset
