+++ docs/Examples/titanium_medium.m
@@ -2,4 +2,4 @@ mtexdata titanium
 grains=calcGrains(ebsd);
 grains=cond_smooth(grains);
 G=gmshGeo(grains);
-mesh(G,'titanium_medium.msh','elementSize',40,'medium',[2000 2000 200],'mediumElementSize',180);
+savegeo(G,'titanium.geo','ElementSize',10,'gradient',5,'ElementType','Tri');
