### 2020-02-19 1e1096f
restore verbosity when meshing
