+++ gmshGeo.m
@@ -554,7 +554,7 @@ classdef gmshGeo
 				end
 			end			
 			tmp_file=obj.savegeo(tempname,varargin{:});	%	Save the geometry into a temp file
-			str=sprintf('"%s" "%s" -o "%s" -v 0 -3',path_to_gmsh,tmp_file,outputFilePath);
+			str=sprintf('"%s" "%s" -o "%s" -v 4 -3',path_to_gmsh,tmp_file,outputFilePath);
 			system(str);
 			delete(tmp_file)	%	delete temp file
 		end
