### 2020-02-19 7fa7fc6
Create CONTRIBUTING.md
