+++ CONTRIBUTING.md
@@ -1,0 +1,23 @@
+# Contribute to this work
+MTEX2Gmsh is developped under the [MIT licence](LICENSE). Thus, you are welcome to contribute in any way.
+This file gives the guidelines to whom may be interested in contributing. 
+
+## Ressources
+Before requesting help or trying to contribute, please first check at the following ressources:
+- [Online documentation](https://doriandepriester.github.io/MTEX2Gmsh/html/index.html),
+- [Issue tracker](/issues).
+
+## Bugs
+Use the Issue tracker for reporting a bug or requesting help. Please, keep in mind that I ([@DorianDepriester](https://github.com/DorianDepriester)) am currently the only 
+maintainer of this toolbox. 
+Thus, I may take a while for considering your issues, depending on external factors such as other works, vacations etc.
+
+## Testing
+If you make any change in the core code, run the examples scripts (in [``/docs/Examples``](/docs/Examples)) to check that everything runs fine with a large variety of geometries.
+
+## Submit changes
+If your contribution passes the aforementioned tests and if you want to share it, create a pull request (see [here](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request) for details).
+Again, do not feel offended if I do not respond quickly.
+
+## Code of conduct
+The Contributor Covenant V2 applies here. This code of conduct is available at https://www.contributor-covenant.org/version/2/0/code_of_conduct/
