### 2020-02-19 f2709d2
Add reference to CONTRIBUTING.md
