+++ README.md
@@ -12,6 +12,9 @@ This toolbox has been designed for MATLAB R2013b, but it may work or newer versi
 ## Documentation and examples
 Visit the corresponding [site](https://doriandepriester.github.io/MTEX2Gmsh/) to see examples and [full documentation](https://doriandepriester.github.io/MTEX2Gmsh/html/index.html). Alternatively, you can check out the [``docs/Examples``](https://github.com/DorianDepriester/MTEX2Gmsh/tree/master/docs/Examples) folder.
 
+## Unit test
+The aforementioned examples can be easily reproduced. In addition, the reader can check out the reproductibility of minimal example on [Code Ocean](https://codeocean.com/capsule/8758800/tree/v1).
+
 ## Reference
 The corresponding paper is still under review. In the meantime, please cite this code using the related DOI:
 [![DOI](https://zenodo.org/badge/137471547.svg)](https://zenodo.org/badge/latestdoi/137471547)
