### 2020-02-19 DorianDepriester/MTEX2Gmsh#1 @DorianDepriester
Thank you @streeve for your quick review. All the points you mentioned have been taken into consideration.

Just to mention, the unit test are not that easy to implement on MATLAB. As far as I know, [Code Ocean](https://codeocean.com/) is the easiest way to provide offline test for reproducibility. The link to the related _Capsule_ has been added in ``README.md``.

Considering the API docs, I am not sure to understand its meaning in the MATLAB scripting context. All functions and methods are provided along with help sections (in addition to documentation). For instance, juste type

    help gmshGeo

to get help about the ``gmshGeo`` constructor.
