### 2020-02-19 f5b261e
Add hack for linux users
