### 2020-02-20 DorianDepriester/MTEX2Gmsh#1 @streeve
Great, thanks

The Code Ocean online environment is quite nice. My only objection is that (as far as I can tell) it only tests one case. Something like https://github.com/vitorsr/SIHR/blob/master/utils/automated_testing.m could work
