### 2020-02-20 DorianDepriester/MTEX2Gmsh#1 @streeve
For the API docs, a section just mentioning both `help` (thank you, I haven't actively used matlab in a while) and `methods` to make it easy to explore all options in the class
