### 2020-02-20 DorianDepriester/MTEX2Gmsh#2 @streeve
matlab warnings
This is an extremely minor issue: there are a handful of suggestions from the matlab warnings (extra semicolons, unnecessary variables) in the gmshGeo class

openjournals/joss-reviews#2094
