### 2020-02-20 DorianDepriester/MTEX2Gmsh#3 @streeve
Paper revisions
- [x] The paper may be more accessible to a general (materials science) audience if the first two sentences are flipped: grain morphology and texture are greatly important for materials behavior, and micromechanics is of great use to that end.
- [x] In paragraph 2 please note these are experimental techniques
- [x] Are you limited to a subset of phases based on implementation, or is it generalized? More specifically, you mention a cubic "criterion", but I do not understand the extension to the current work.
- [x] As I understand, your work builds on the previous by expanding to (contingent on the previous question) any phases/orientations by combining the two underlying codes - please make this explicit
- [x] Within the figure caption you recite mtex, but not gmsh

openjournals/joss-reviews#2094
