### 2020-02-24 5843361
Update paper.md
Fixed issue #3
