### 2020-02-24 2027e9c
Add hints for help functions
Fix 6th item in Issue #1
