### 2020-02-24 DorianDepriester/MTEX2Gmsh#3 @DorianDepriester
Thank you @streeve
All the items you mentioned have been taken into consideration. Please, check out the [new proof](https://github.com/openjournals/joss-reviews/issues/2094#issuecomment-590469247) and tell me if it fits with your taste.
