### 2020-02-26 DorianDepriester/MTEX2Gmsh#2 @DorianDepriester
The code analyzer shows neither error nor warning in all .m files (tested on Matlab 2013b and 2016a). Which editor do you use?
