### 2020-02-27 DorianDepriester/MTEX2Gmsh#1 @DorianDepriester
I have added a file named [unitTest](https://github.com/DorianDepriester/MTEX2Gmsh/blob/master/docs/Examples/unitTest.m). It sums up all the examples given in the eponymous folder, plus it plots the geometry in each case.
