### 2020-02-27 DorianDepriester/MTEX2Gmsh#1 @DorianDepriester
Alternatively, I have update the code on code Ocean, following your recommendations (that is, provide multiple examples).

Tell me if if need to keep the aforementioned ``unitTest.m`` or/and the reference to the capsule.
