### 2020-02-28 DorianDepriester/MTEX2Gmsh#1 @streeve
I think `unitTest.m` is exactly what you want: it runs the main examples, is easy to extend, and you can run on Code Ocean or locally
