### 2020-03-02 DorianDepriester/MTEX2Gmsh#2 @DorianDepriester
The warnings you've mentioned are now fixed (tested on MATLAB 2018a).
