### 2020-03-02 #2094 @streeve
@meg-simula @DorianDepriester

I've completed the checklist and closed all the issues with my suggestions. This is a very nice code, combining the strengths of underlying tools. I'm no expert in this type of simulation, but the docs and examples make it clear what I could do with it.

One more note: I did not test the Abaqus portion (I do not have access)
