### 2020-05-27 6e8f7b4
Fix "Array indices must be positive blablabla" with older versions of MATLAB
