### 2020-05-27 cc63611
Warning when the number of ambiguous paths is large
