### 2020-06-30 DorianDepriester/MTEX2Gmsh#5 @ralfHielscher
Paper review
Hi Dorian,

I went through the code and found it very useful and impressive. I think there are some places for improvement.
I will list them here but I don't think that they are showstopper for the review process.

1. A central point seems to be the reordering of the boundary segments . Here you use an algorithm based on Euler Tours. MTEX itself contains already some Euler Tour algorithms in `mtex/tools/graph_tools` which might be more efficient. Another idea can be found in the last paragraph of `grain2d/subSet`. Here the boundary segments are ordered without the use of any Euler Tours command by making use of the fact that `grain2d.poly` contains already a ordered list of vertices.

2. For me it would make a lot of sense to include the ordering directly into MTEX. Currently, the grainboundary segments are not ordered at all (unless there is only a single grains). The ordering you suggest seems very reasonable and could easily be the default ordering in MTEX.

3. The spline interpolation seems to be used only for plotting. This is not so clear from the documentation. Also, wouldn't it make more sense to use some buildin spline function like `spline` instead of a custom one.

4. Smoothing the boundary while keeping triple points and points at the outer boundary fixed should actually be an option within MTEX :) Same is true for the simplification of the geometry,

5. When publishing the help files it is useful to set
```
setMTEXpref('generatingHelpMode',true);
```
as it avoids some output  artifacts.

6. I'm not completely convinced about having an extra class for the grains. Wouldn't it make more sense to inherit from `grain2d`? When I understood this correctly the main difference to the class `grain2d` is that the boundary segments are sorted and the `gmshGeo` keeps track of this order. I'm correct?
