### 2020-07-01 #2094 @DorianDepriester
Hello there,
Just to mention that @ralfHielscher has kindly reviewed the submission.
https://github.com/DorianDepriester/MTEX2Gmsh/issues/5#issue-648501620
I am doing my best to promptly answer each bullet point.
