### 2020-07-02 d4fb03b
Fixed issue #5
