### 2020-07-02 DorianDepriester/MTEX2Gmsh#5 @DorianDepriester
Thank you @ralfHielscher for your report. Here are the answers to the points you have raised:

1. I am aware that my own EulerPath is not very efficient (very slow compared to the MTEX's EulerCycles). Still, under some rare circumstances, the EulerCycles() fails to walk down the path as I wish. See for instance [file attached](https://github.com/DorianDepriester/MTEX2Gmsh/files/4858367/Steel.zip): because of the self 'touching' grain boundary, the path is ambiguous. This situation generally occurs because of outliers, so it often can be avoided by cleaning the data (e.g. w/o 1-dot-size grains); but sometimes outlier removal do not fix this.

2. Considering the point above, if the EulerCycles roughly complies with my needs in a future version of MTEX, I would be glad to use it instead of my DIY EulerPath.

3. The Bspline approximation is indeed only used in MATLAB for plotting. But it is called in Gmsh before meshing (as a built-in feature). As a result, the implementation of the Bspline function is required here to ensure that the geometry plotted in MATLAB is consistent with that read by Gmsh before meshing.

4. I completely agree with that point. Also note that higher-order vertex can be reached (quadruple points and so on), but they are not listed in the grain2d property `triplePoints`.

5. I have added `setMTEXpref('generatingHelpMode',true);` at the beginning of each file in /doc. See commit d4fb03b628f784d7ce0f2e5a8aeb5b5c93c58cba.

6. I do agree that subclassing grain2d could be clearer, and this could be done in a future release. Actually, the data in the gmshGeo are stored in a very different way; so, for somehow historical reasons, the gmshGeo class has been designed from scratch.
