### 2020-07-24 DorianDepriester/MTEX2Gmsh#5 @ralfHielscher
Hi Dorian,

thank you for your responses. I agree with you in all points. Especially when it comes to the sorting of the boundary segments this is something I still need to do in MTEX.

Congratulations to this piece of software.

Ralf.
