### 2020-07-29 #2094 @DorianDepriester
I have answered all the point raised by @ralfHielscher in the issue [mentioned above](https://github.com/DorianDepriester/MTEX2Gmsh/issues/5#issue-648501620). He seems satisfied for he has closed the issue.
