### 2020-08-20 #2094 @meg-simula
@DorianDepriester Thanks for your patience with this submission.

I've gone through the paper, and ask that you please consider the following suggestions:

- [ ] Include a clear Statement of Need that illustrates the research purpose of the software.
- [ ] Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.
- [ ] Acknowledge any financial support.
- [ ] This sentence is hard to understand: "As part of characterization techniques, Orientation Imaging Microscopy (OIM), usually made from Electron Backscatter Diffraction (EBSD) data, is now widely used in the literature for investigating the grain morphology and local crystal orientations in crystalline materials." Rephrase.
- [ ] Similarly, rephrase: "Nevertheless, because of the regular structure, the GBs can not be smoothly described"
- [ ] Replace  "theory -based" by "theory-based".
