### 2020-08-20 #2094 @DorianDepriester
@meg-simula, here are the answers to all the points you have raised:

- [x] A clear Statement of Need has been added (see https://github.com/DorianDepriester/MTEX2Gmsh/commit/afd5ebbe34fb63173765c18be54a48dadc0aa6ac )
- [x] There is no representative set of past or ongoing research projects using the software and nor recent scholarly publications enabled by it.
- [x] This project is not financially supported, thus there is no need for any acknowledgement
- [x] The sentence has been rephrased (see https://github.com/DorianDepriester/MTEX2Gmsh/commit/d251c90a6896240a192db27394398ea974ed55b7)
- [x] The sentence has been rephrased too (see https://github.com/DorianDepriester/MTEX2Gmsh/commit/256cec593c150e32978c4808cf9f111626beac48)
- [x] The typo has been fixed (see https://github.com/DorianDepriester/MTEX2Gmsh/commit/3a23e22a49c765400f12ecb1531394e2d2d26fe7)
